Move common test code to cluster_utils.rs
Extract the cluster fixture, counting proxies, and the scoped-client and discovery helpers shared by the topology tests into cluster_utils.rs,
so the new test files can reuse them instead of importing from load_balancing_tests.rs
