+++ tests/load_balancing/load_balancing_tests.rs
@@ -1,11 +1,9 @@
 use crate::ccm_wrapper::ccm::*;
 use crate::ccm_wrapper::cluster::*;
-use crate::ccm_wrapper::topology_spec::*;
-use crate::load_balancing::proxy;
+use crate::load_balancing::cluster_utils::*;
 use crate::load_balancing::scope_utils;
 
 use alternator_driver::AlternatorClient;
-use alternator_driver::AlternatorConfig;
 use alternator_driver::RoutingScope;
 use alternator_driver::keyrouting::affinity_config::{
     KeyRouteAffinityConfig, KeyRouteAffinityType,
@@ -82,213 +80,29 @@ fn redirect_target_node(cluster: &Cluster) -> Node {
     cluster.datacenters()[0].racks()[0].nodes()[0].clone()
 }
 
-// Struct for counting requests made to the proxy. GETs, POSTs, and describe_tables are counted separately.
-// GETs are service discovery calls, POSTs are the actual calls to DB, and describe_table calls
-// are from PartitionKeyResolver.
-#[derive(Debug)]
-pub(crate) struct NodeCounter {
-    posts: AtomicUsize,
-    gets: AtomicUsize,
-    describe_tables: AtomicUsize,
-}
-
-impl NodeCounter {
-    fn new() -> Self {
-        Self {
-            posts: AtomicUsize::new(0),
-            gets: AtomicUsize::new(0),
-            describe_tables: AtomicUsize::new(0),
-        }
-    }
-
-    pub(crate) fn posts(&self) -> usize {
-        self.posts.load(Ordering::Relaxed)
-    }
-
-    fn reset(&self) {
-        self.posts.store(0, Ordering::Relaxed);
-        self.gets.store(0, Ordering::Relaxed);
-        self.describe_tables.store(0, Ordering::Relaxed);
-    }
-}
-
-pub(crate) async fn start_counting_proxy(
-    listen_addr: String,
-    connect_addr: String,
-    request_counter: Arc<NodeCounter>,
-) {
-    let proxy = proxy::Proxy::start(
-        listen_addr,
-        connect_addr,
-        move |req, send| {
-            let node_counter = Arc::clone(&request_counter);
-            async move {
-                {
-                    let is_describe_table = req
-                        .headers()
-                        .get("x-amz-target")
-                        .is_some_and(|h| h == "DynamoDB_20120810.DescribeTable");
-
-                    match *req.method() {
-                        Method::POST => {
-                            if is_describe_table {
-                                node_counter.describe_tables.fetch_add(1, Ordering::Relaxed);
-                            } else {
-                                node_counter.posts.fetch_add(1, Ordering::Relaxed);
-                            }
-                        }
-                        Method::GET => {
-                            node_counter.gets.fetch_add(1, Ordering::Relaxed);
-                        }
-                        _ => {}
-                    };
-                }
-                proxy::forward_on_request(req, send).await
-            }
-        },
-        None,
-        None,
-    )
-    .await;
-
-    // Avoid a dead-code warning.
-    let _ = proxy.address();
-
-    // We deliberately detach the proxy task here.
-    // Each test has its own Tokio runtime, so dropping the runtime will abort
-    // this task and cleanly release the listener and connection resources.
-    // Only one test at a time can use the cluster, so old proxies will be dropped before new ones are created.
-    tokio::spawn(async move {
-        proxy.run().await;
-    });
-}
-
 // This proxy is used in tests where we check if calls are made to a node that is down.
 // Redirecting calls to a different node allows us to count calls to the stopped node,
 // while also allowing the client to use it for discovery without connection errors.
-pub(crate) async fn start_redirecting_proxy(
-    from: &Node,
-    to: &Node,
-    request_counter: Arc<NodeCounter>,
-) {
+async fn start_redirecting_proxy(from: &Node, to: &Node, request_counter: Arc<NodeCounter>) {
     let listen_addr = format!("{}:{}", from.ip.clone(), from.alternator_port);
     let connect_addr = format!("{}:{}", to.ip.clone(), ALTERNATOR_PORT);
     start_counting_proxy(listen_addr, connect_addr, request_counter).await;
 }
 
-// This is the proxy that calls to the node go through. Used to count calls and ensure that client calls the correct nodes.
-pub(crate) async fn start_proxy_on_node(
-    node: Node,
-    proxy_port: u16,
-    request_counter: Arc<NodeCounter>,
-) {
-    let listen_addr = format!("{}:{}", node.ip, proxy_port);
-    let connect_addr = format!("{}:{}", node.ip, ALTERNATOR_PORT);
-    start_counting_proxy(listen_addr, connect_addr, request_counter).await;
-}
-
-pub(crate) async fn start_proxies(
-    cluster: &mut Cluster,
-    proxy_port: u16,
-    request_counter: &RequestCounter,
-) {
-    cluster.update_all_nodes_port(proxy_port);
-    let counter = &request_counter.counter;
-    for node in cluster.nodes() {
-        if node.is_up {
-            let node_counter = Arc::clone(counter.get(&node.ip).unwrap());
-            start_proxy_on_node(node.clone(), proxy_port, node_counter).await;
-        }
-    }
-}
-
-pub(crate) async fn start_redirecting_proxies(
+async fn start_redirecting_proxies(
     cluster: &mut Cluster,
     proxy_port: u16,
     request_counter: &RequestCounter,
 ) {
     cluster.update_all_nodes_port(proxy_port);
     let to = &redirect_target_node(cluster);
-    let counter = &request_counter.counter;
     for node in cluster.nodes() {
-        let node_counter = Arc::clone(counter.get(&node.ip).unwrap());
+        let node_counter = request_counter.get(&node.ip);
         start_redirecting_proxy(node, to, node_counter).await;
     }
 }
 
-// Struct with a hashmap underneath for holding and monitoring the counters for multiple nodes.
-#[derive(Debug)]
-pub(crate) struct RequestCounter {
-    counter: HashMap<String, Arc<NodeCounter>>,
-}
-
-impl RequestCounter {
-    pub(crate) fn new() -> Self {
-        Self {
-            counter: HashMap::new(),
-        }
-    }
-
-    pub(crate) fn from_cluster(cluster: &Cluster) -> Self {
-        let counter = cluster
-            .nodes()
-            .iter()
-            .map(|node| (node.ip.clone(), Arc::new(NodeCounter::new())))
-            .collect();
-        Self { counter }
-    }
-
-    pub(crate) fn get(&self, ip: &str) -> Arc<NodeCounter> {
-        Arc::clone(self.counter.get(ip).unwrap())
-    }
-
-    pub(crate) fn add(&mut self, ip: String) {
-        self.counter.insert(ip, Arc::new(NodeCounter::new()));
-    }
-
-    pub(crate) fn reset(&self) {
-        for c in self.counter.values() {
-            c.reset();
-        }
-    }
-
-    pub(crate) fn total_posts(&self) -> usize {
-        self.counter
-            .values()
-            .map(|c| c.posts.load(Ordering::Relaxed))
-            .sum()
-    }
-
-    pub(crate) fn get_posts_to_ips(&self, ips: &[&str]) -> usize {
-        ips.iter()
-            .map(|ip| self.counter.get(*ip).unwrap().posts.load(Ordering::Relaxed))
-            .sum()
-    }
-
-    pub(crate) fn get_posts_to_other_ips(&self, ips: &[&str]) -> usize {
-        self.counter
-            .iter()
-            .filter(|(ip, _)| !ips.contains(&ip.as_str()))
-            .map(|(_, c)| c.posts.load(Ordering::Relaxed))
-            .sum()
-    }
-
-    pub(crate) fn total_describe_tables(&self) -> usize {
-        self.counter
-            .values()
-            .map(|c| c.describe_tables.load(Ordering::Relaxed))
-            .sum()
-    }
-}
-
-// Make calls without caring about the result, used to count where calls are directed.
-async fn make_n_calls(client: &AlternatorClient, n: usize) {
-    for _ in 0..n {
-        let _ = client.list_tables().send().await;
-    }
-}
-
-pub(crate) async fn create_table(client: &AlternatorClient, table_name: &str) {
+async fn create_table(client: &AlternatorClient, table_name: &str) {
     client
         .create_table()
         .table_name(table_name)
@@ -312,7 +126,7 @@ pub(crate) async fn create_table(client: &AlternatorClient, table_name: &str) {
         .unwrap();
 }
 
-pub(crate) async fn put_item(client: &AlternatorClient, table_name: &str, item: &str) {
+async fn put_item(client: &AlternatorClient, table_name: &str, item: &str) {
     let _ = client
         .put_item()
         .table_name(table_name)
@@ -324,7 +138,7 @@ pub(crate) async fn put_item(client: &AlternatorClient, table_name: &str, item: 
         .await;
 }
 
-pub(crate) async fn delete_item(client: &AlternatorClient, table_name: &str, item: &str) {
+async fn delete_item(client: &AlternatorClient, table_name: &str, item: &str) {
     let _ = client
         .delete_item()
         .table_name(table_name)
@@ -336,12 +150,7 @@ pub(crate) async fn delete_item(client: &AlternatorClient, table_name: &str, ite
         .await;
 }
 
-pub(crate) async fn update_item(
-    client: &AlternatorClient,
-    table_name: &str,
-    item: &str,
-    new: &str,
-) {
+async fn update_item(client: &AlternatorClient, table_name: &str, item: &str, new: &str) {
     let _ = client
         .update_item()
         .table_name(table_name)
@@ -358,55 +167,13 @@ pub(crate) async fn update_item(
         .await;
 }
 
-// Create a basic client with scope.
-fn create_client_with_scope(cluster: &Cluster, scope: RoutingScope) -> AlternatorClient {
-    AlternatorClient::from_conf(
-        AlternatorConfig::builder()
-            .credentials_provider(
-                aws_sdk_dynamodb::config::Credentials::for_tests_with_session_token(),
-            )
-            .region(aws_sdk_dynamodb::config::Region::new("eu-central-1"))
-            .behavior_version_latest()
-            .endpoint_url(default_endpoint_url(cluster))
-            .routing_scope(scope)
-            .build(),
-    )
-}
-
-// Poll until the client's live nodes match the given IPs, or timeout.
-async fn wait_until_live_nodes_match(client: &AlternatorClient, ips: Vec<&str>) {
-    let live_nodes = client.config().live_nodes().unwrap().clone();
-    tokio::time::timeout(POLLING_TIMEOUT, async {
-        loop {
-            let nodes = live_nodes.get_live_nodes();
-            let node_ips: Vec<&str> = nodes.iter().map(|url| url.host_str().unwrap()).collect();
-            if node_ips.len() == ips.len() && node_ips.iter().all(|ip| ips.contains(ip)) {
-                break;
-            }
-            tokio::time::sleep(POLLING_INTERVAL).await;
-        }
-    })
-    .await
-    .unwrap_or_else(|_| {
-        panic!(
-            "failed to update nodes\nexpected nodes {:?}, timed out after {:?}",
-            ips, POLLING_TIMEOUT
-        )
-    });
-}
-
 fn create_client_with_scope_and_affinity(
     cluster: &Cluster,
     scope: RoutingScope,
     affinity_config: KeyRouteAffinityConfig,
 ) -> AlternatorClient {
     AlternatorClient::from_conf(
-        AlternatorConfig::builder()
-            .credentials_provider(
-                aws_sdk_dynamodb::config::Credentials::for_tests_with_session_token(),
-            )
-            .region(aws_sdk_dynamodb::config::Region::new("eu-central-1"))
-            .behavior_version_latest()
+        minimal_builder()
             .endpoint_url(default_endpoint_url(cluster))
             .routing_scope(scope)
             .key_route_affinity(affinity_config)
